# swagger.yaml
openapi: 3.0.3
...
info:
  title: Laundry Service API
  description: |
    ## 🧺 Tổng quan
    API cho nền tảng dịch vụ Giặt ủi trực tuyến. Hệ thống hỗ trợ:
    - **Khách hàng (Customer)**: Đặt đơn giặt, theo dõi trạng thái, thanh toán
    - **Nhân viên (Staff)**: Xử lý đơn hàng, cập nhật trạng thái, upload ảnh
    - **Quản trị viên (Admin)**: Toàn quyền quản lý hệ thống

    ## 🔐 Xác thực
    Hệ thống hỗ trợ **2 luồng đăng nhập**:

    ### Luồng 1: Email + Password (dành cho Admin & Staff)
    ```
    POST /auth/register → POST /auth/login → nhận { accessToken }
    ```

    ### Luồng 2: SĐT + OTP (dành cho Customer — không cần password)
    ```
    POST /auth/send-otp → POST /auth/verify-otp → nhận { accessToken }
    ```

    ### Cách truyền Token lên Header
    ```
    Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    ```

    ### Refresh Token
    - Được lưu trong **HttpOnly Cookie** (`refreshToken`), browser tự gửi kèm
    - Khi access token hết hạn, gọi `POST /auth/refresh-token` để cấp lại

    ## 📦 Luồng trạng thái đơn hàng
    ```
    received → washing → drying → delivering → completed
        ↓          ↓        ↓           ↓
    cancelled  cancelled cancelled  cancelled
    ```
    *(Customer chỉ hủy được khi status = received. Admin hủy được mọi lúc trừ completed)*

    ## 📝 Cấu trúc Response chuẩn
    Mọi response đều theo format:
    ```json
    {
      "success": true,
      "message": "Mô tả kết quả",
      "data": { ... },
...
    }
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token. Lấy từ response của `/auth/login` hoặc `/auth/verify-otp`.
        Truyền vào header: `Authorization: Bearer <token>`

  # ═══════════════════════════════════════════════════════════════════════════
  #  SCHEMAS (Data Models / Types)
  #  FE có thể dùng phần này để tạo TypeScript Interface/Type
  # ═══════════════════════════════════════════════════════════════════════════
  schemas:
...
    User:
      type: object
      description: Thông tin người dùng (không bao gồm password)
      properties:
        _id:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d1"
          description: MongoDB ObjectId
        name:
          type: string
          example: "Nguyễn Văn An"
          description: Họ và tên
        email:
          type: string
          format: email
          example: "nguyenvanan@gmail.com"
          description: Email (optional — chỉ có với luồng email+password)
          nullable: true
        phone:
          type: string
          example: "0912345678"
          description: Số điện thoại (optional — chỉ có với luồng OTP)
          nullable: true
        address:
          type: string
          example: "123 Lê Lợi, Quận 1, TP.HCM"
          nullable: true
        role:
          type: string
          enum: [customer, staff, admin]
          example: "customer"
          description: Vai trò người dùng
        isActive:
          type: boolean
          example: true
        isPhoneVerified:
          type: boolean
          example: false
          description: SĐT đã được xác thực qua OTP chưa
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-06-20T08:00:00.000Z"

    # ─── SERVICE (Gói dịch vụ) ────────────────────────────────────────────────
...
    SendOtpRequest:
      type: object
      required: [phone]
      description: Body để gửi mã OTP tới số điện thoại
      properties:
        phone:
          type: string
          example: "0912345678"
          description: Số điện thoại VN (định dạng 10 số, bắt đầu bằng 0)
    VerifyOtpRequest:
      type: object
      required: [phone, otp]
      description: Body để xác thực mã OTP và lấy JWT token
      properties:
        phone:
          type: string
          example: "0912345678"
        otp:
          type: string
          example: "123456"
          minLength: 6
          maxLength: 6
          description: Mã OTP 6 chữ số nhận được qua SMS
    UpdateProfileRequest:
      type: object
      description: Body để cập nhật thông tin cá nhân
      properties:
        name:
          type: string
          example: "Nguyễn Văn An (updated)"
        phone:
          type: string
          example: "0987654321"
        address:
          type: string
          example: "456 Nguyễn Trãi, Quận 5, TP.HCM"
...
    VerifyOtpResponse:
      type: object
      description: Response sau khi xác thực OTP thành công
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Đăng nhập thành công!"
        data:
          type: object
          properties:
            accessToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            user:
              $ref: '#/components/schemas/User'
            isNewUser:
              type: boolean
              example: false
              description: true nếu đây là lần đăng ký đầu tiên, false nếu đăng nhập lại
...
    ErrorResponse:
      type: object
      description: Cấu trúc lỗi chung
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Mô tả lỗi cụ thể"
        data:
          nullable: true
          example: null
    ValidationErrorResponse:
      type: object
      description: Lỗi validation (400) — khi dữ liệu đầu vào không hợp lệ
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Dữ liệu không hợp lệ"
        errors:
          type: array
          items:
            type: string
          example: ["Tên dịch vụ không được để trống", "Giá tiền không được âm"]
        data:
          nullable: true
          example: null

  # ─── REUSABLE PARAMETERS ──────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Token không hợp lệ hoặc chưa đăng nhập
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Bạn chưa đăng nhập. Vui lòng đăng nhập để tiếp tục."
            data: null
...
    BadRequest:
      description: Dữ liệu đầu vào không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
security:
  - BearerAuth: []

# ═══════════════════════════════════════════════════════════════════════════════
#  TAGS — Nhóm các endpoint
# ═══════════════════════════════════════════════════════════════════════════════
...
paths:
...
  /auth/send-otp:
    post:
      tags: [Auth]
      summary: Gửi mã OTP tới số điện thoại (Bước 1 — Luồng OTP)
      description: |
        Bước 1 trong luồng đăng nhập bằng SĐT+OTP dành cho Customer.

        - Nếu SĐT chưa tồn tại trong hệ thống → OTP dùng để đăng ký tài khoản mới
        - Nếu SĐT đã tồn tại → OTP dùng để đăng nhập

        **Rate limit:** Tối đa 5 lần/15 phút/IP

        **OTP hết hạn sau:** 5 phút
      operationId: sendOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpRequest'
            example:
              phone: "0912345678"
      responses:
        '200':
          description: Gửi OTP thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Đã gửi mã OTP tới số 0912345678"
                  data:
                    type: object
                    properties:
                      expiresInMinutes:
                        type: number
                        example: 5
                        description: OTP hết hạn sau bao nhiêu phút
        '400':
          description: Thiếu số điện thoại hoặc định dạng không đúng
          $ref: '#/components/responses/BadRequest'
        '429':
          description: Vượt quá rate limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Quá nhiều yêu cầu. Vui lòng thử lại sau 15 phút."
                data: null
  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: Xác thực mã OTP và lấy JWT Token (Bước 2 — Luồng OTP)
      description: |
        Bước 2 trong luồng đăng nhập bằng SĐT+OTP.

        - Nếu là user mới → tự động tạo tài khoản và đăng nhập, trả về `isNewUser: true`
        - Nếu là user cũ → đăng nhập, trả về `isNewUser: false`

        **Rate limit:** Tối đa 10 lần/15 phút/IP

        **Refresh Token** được set tự động vào HttpOnly Cookie.
      operationId: verifyOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
            example:
              phone: "0912345678"
              otp: "123456"
      responses:
        '200':
          description: Đăng nhập thành công (user cũ)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyOtpResponse'
        '201':
          description: Đăng ký thành công (user mới)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyOtpResponse'
              example:
                success: true
                message: "Đăng ký thành công! Chào mừng bạn đến với LaundryApp."
                data:
                  accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    _id: "64f1a2b3c4d5e6f7a8b9c0d1"
                    name: ""
                    phone: "0912345678"
                    role: "customer"
                    isPhoneVerified: true
                    isActive: true
                  isNewUser: true
        '400':
          description: OTP không đúng hoặc đã hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Mã OTP không hợp lệ hoặc đã hết hạn"
                data: null
        '429':
          description: Vượt quá rate limit
  /auth/refresh-token:
    post:
      tags: [Auth]
      summary: Cấp lại Access Token mới bằng Refresh Token
      description: |
        FE gọi endpoint này khi nhận được lỗi `401 Unauthorized` (access token hết hạn).

        **Cơ chế hoạt động:**
        - `refreshToken` được gửi tự động qua HttpOnly Cookie (browser tự xử lý)
        - Server xác thực cookie và cấp lại `accessToken` mới

        > ⚠️ Không cần truyền gì trong body. Cookie được gửi tự động.
      operationId: refreshToken
      security: []
      responses:
        '200':
          description: Cấp lại token thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Cấp lại access token thành công"
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Refresh token không hợp lệ hoặc đã hết hạn — yêu cầu đăng nhập lại
          $ref: '#/components/responses/Unauthorized'
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Lấy thông tin tài khoản đang đăng nhập
      description: |
        Trả về đầy đủ thông tin của user đang đăng nhập (dựa vào JWT token trong header).
        Tương đương với `GET /users/profile`.
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lấy thông tin thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Lấy thông tin thành công"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Auth]
      summary: Cập nhật thông tin cá nhân
      description: |
        Cập nhật thông tin cá nhân của user đang đăng nhập.
        Chỉ cần truyền các field muốn thay đổi.
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              name: "Nguyễn Văn An"
              address: "789 Trần Hưng Đạo, Quận 1, TP.HCM"
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Cập nhật thành công"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─────────────────────────────────────────────────────────────────────────────
  #  SERVICES ENDPOINTS  (Base: /api/services)
  # ─────────────────────────────────────────────────────────────────────────────
...